LayoutTests/imported/w3c/web-platform-tests/css/css-values/attr-security-if-expected.txt


FAIL '--x: image-set(if(style(--true): attr(data-foo);))' with data-foo="https://does-not-exist.test/404.png" assert_equals: expected "image-set(\"https://does-not-exist.test/404.png\")" but got "image-set(if(style(--true): \"https://does-not-exist.test/404.png\";))"
PASS 'background-image: image-set(if(style(--true): attr(data-foo);))' with data-foo="https://does-not-exist.test/404.png"
PASS 'background-image: image-set(
                if(style(--true): url(https://does-not-exist-2.test/404.png);
                   else: attr(data-foo);))' with data-foo="https://does-not-exist-2.test/404.png"
PASS 'background-image: image-set(
                if(style(--some-string): url(https://does-not-exist.test/404.png);))' with data-foo="https://does-not-exist.test/404.png"
PASS 'background-image: image-set(
                if(style(--condition-val: attr(data-foo type(*))): url(https://does-not-exist.test/404.png);))' with data-foo="3"
PASS 'background-image: image-set(
                if(style(--condition-val: attr(data-foo type(*))): url(https://does-not-exist.test/404.png);
                   style(--true): url(https://does-not-exist.test/404.png);
                   else: url(https://does-not-exist.test/404.png);))' with data-foo="1"
PASS 'background-image: image-set(if(style(--true): url(https://does-not-exist.test/404.png);
                            style(--condition-val): url(https://does-not-exist.test/404.png);
                            else: url(https://does-not-exist.test/404.png);))' with data-foo="attr(data-foo type(*))"
PASS 'background-image: image-set(
                if(style(--condition-val: if(style(--true): attr(data-foo type(*));)): url(https://does-not-exist.test/404.png);))' with data-foo="3"
FAIL '--x: image-set(if(style(--condition-val: if(style(--true): attr(data-foo type(*));)): url(https://does-not-exist.test/404.png);))' with data-foo="3" assert_equals: expected "image-set(url(https://does-not-exist.test/404.png))" but got "image-set(if(style(--condition-val: if(style(--true): 3;)): url(https://does-not-exist.test/404.png);))"
FAIL '--x: image-set(if(style(--condition-val >= attr(data-foo type(*))): url(https://does-not-exist.test/404.png);))' with data-foo="3" assert_equals: expected "image-set(url(https://does-not-exist.test/404.png))" but got "image-set(if(style(--condition-val >= 3): url(https://does-not-exist.test/404.png);))"
PASS 'background-image: image-set(
                if(style(--condition-val >= attr(data-foo type(*))): url(https://does-not-exist.test/404.png);))' with data-foo="3"
PASS 'background-image: image-set(
                if(style(--condition-val < attr(data-foo type(*))): url(https://does-not-exist.test/404.png);))' with data-foo="3"
PASS 'background-image: image-set(
                if(style(--str < attr(data-foo type(*))): url(https://does-not-exist.test/404.png);))' with data-foo="3"
PASS 'background-image: image-set(
                if(style(--condition-val < attr(data-foo type(*))): url(https://does-not-exist.test/404.png);))' with data-foo="text"
